requirements.txt
torch
pytorch
pytorch-cuda=12.1
torchvision
torchmetrics
torchaudio
torchtext
tensorboard
pyyaml
tqdm
numpy
datasets
plotly
matplotlib
kaleido